fusion/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
coverage==7.3.2
dj-database-url==2.1.0
dj-static==0.0.6
Django==4.2.6
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-stdimage==6.0.1
gunicorn==21.2.0
model-mommy==2.0.0
mysql==0.0.3
mysqlclient==2.2.0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
PyMySQL==1.1.0
soupsieve==2.5
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
